secp256k1: Split funcs for crypto/elliptic iface.
This is the first of a series of commits that intends to make the
internals of the package entirely independent of the crypto/elliptic and
crypto/ecdsa stdlib interfaces.  The package is already highly specific
to secp256k1 in many areas, but it still has a lot of cruft from the
early days when it was a generic elliptic curve package.

This ultimately will allow a lot more optimization specific to secp256k1
to be made once it can break free of the restrictive standard lib
crypto/elliptic interface.

However, in an effort to maintain compatibility with the standard lib,
the necessary interfaces and types will still be provided, but the goal
is to provide them as a wrapper as opposed to implementing everything
internally in terms of them.

To that end, this simply moves the type and specific functions that are
required to implement the interface into a separate file named
ellipticadaptor.go without making any other changes outside of some
comment updates and removal of unnecessary funcs.
