+++ dcrec/secp256k1/btcec.go
@@ -1,5 +1,5 @@
 // Copyright 2010 The Go Authors. All rights reserved.
-// Copyright (c) 2015-2019 The Decred developers
+// Copyright (c) 2015-2020 The Decred developers
 // Copyright 2011 ThePiachu. All rights reserved.
 // Copyright 2013-2014 The btcsuite developers
 // Use of this source code is governed by an ISC
@@ -8,9 +8,6 @@
 package secp256k1
 
 // References:
-//   [SECG]: Recommended Elliptic Curve Domain Parameters
-//     https://www.secg.org/sec2-v2.pdf
-//
 //   [GECC]: Guide to Elliptic Curve Cryptography (Hankerson, Menezes, Vanstone)
 
 // This package operates, internally, on Jacobian coordinates. For a given
@@ -21,9 +18,7 @@ package secp256k1
 // reverse the transform than to operate in affine coordinates.
 
 import (
-	"crypto/elliptic"
 	"math/big"
-	"sync"
 )
 
 var (
@@ -32,43 +27,6 @@ var (
 	// arithmetic.
 	fieldOne = new(fieldVal).SetInt(1)
 )
-
-// KoblitzCurve supports a koblitz curve implementation that fits the ECC Curve
-// interface from crypto/elliptic.
-type KoblitzCurve struct {
-	*elliptic.CurveParams
-	q *big.Int
-	H int // cofactor of the curve.
-
-	// byteSize is simply the bit size / 8 and is provided for convenience
-	// since it is calculated repeatedly.
-	byteSize int
-
-	// bytePoints
-	bytePoints *[32][256][3]fieldVal
-
-	// The next 6 values are used specifically for endomorphism
-	// optimizations in ScalarMult.
-
-	// lambda must fulfill lambda^3 = 1 mod N where N is the order of G.
-	lambda *big.Int
-
-	// beta must fulfill beta^3 = 1 mod P where P is the prime field of the
-	// curve.
-	beta *fieldVal
-
-	// See the EndomorphismVectors in gensecp256k1.go to see how these are
-	// derived.
-	a1 *big.Int
-	b1 *big.Int
-	a2 *big.Int
-	b2 *big.Int
-}
-
-// Params returns the parameters for the curve.
-func (curve *KoblitzCurve) Params() *elliptic.CurveParams {
-	return curve.CurveParams
-}
 
 // bigAffineToField takes an affine point (x, y) as big integers and converts
 // it to an affine point as field values.
@@ -103,19 +61,6 @@ func (curve *KoblitzCurve) fieldJacobianToBigAffine(x, y, z *fieldVal) (*big.Int
 	x3.SetBytes(x.Bytes()[:])
 	y3.SetBytes(y.Bytes()[:])
 	return x3, y3
-}
-
-// IsOnCurve returns boolean if the point (x,y) is on the curve.
-// Part of the elliptic.Curve interface. This function differs from the
-// crypto/elliptic algorithm since a = 0 not -3.
-func (curve *KoblitzCurve) IsOnCurve(x, y *big.Int) bool {
-	// Convert big ints to field values for faster arithmetic.
-	fx, fy := curve.bigAffineToField(x, y)
-
-	// Elliptic curve equation for secp256k1 is: y^2 = x^3 + 7
-	y2 := new(fieldVal).SquareVal(fy).Normalize()
-	result := new(fieldVal).SquareVal(fx).Mul(fx).AddInt(7).Normalize()
-	return y2.Equals(result)
 }
 
 // addZ1AndZ2EqualsOne adds two Jacobian points that are already known to have
@@ -449,31 +394,6 @@ func (curve *KoblitzCurve) addJacobian(x1, y1, z1, x2, y2, z2, x3, y3, z3 *field
 	curve.addGeneric(x1, y1, z1, x2, y2, z2, x3, y3, z3)
 }
 
-// Add returns the sum of (x1,y1) and (x2,y2). Part of the elliptic.Curve
-// interface.
-func (curve *KoblitzCurve) Add(x1, y1, x2, y2 *big.Int) (*big.Int, *big.Int) {
-	// A point at infinity is the identity according to the group law for
-	// elliptic curve cryptography.  Thus, ∞ + P = P and P + ∞ = P.
-	if x1.Sign() == 0 && y1.Sign() == 0 {
-		return x2, y2
-	}
-	if x2.Sign() == 0 && y2.Sign() == 0 {
-		return x1, y1
-	}
-
-	// Convert the affine coordinates from big integers to field values
-	// and do the point addition in Jacobian projective space.
-	fx1, fy1 := curve.bigAffineToField(x1, y1)
-	fx2, fy2 := curve.bigAffineToField(x2, y2)
-	fx3, fy3, fz3 := new(fieldVal), new(fieldVal), new(fieldVal)
-	fOne := new(fieldVal).SetInt(1)
-	curve.addJacobian(fx1, fy1, fOne, fx2, fy2, fOne, fx3, fy3, fz3)
-
-	// Convert the Jacobian coordinate field values back to affine big
-	// integers.
-	return curve.fieldJacobianToBigAffine(fx3, fy3, fz3)
-}
-
 // doubleZ1EqualsOne performs point doubling on the passed Jacobian point
 // when the point is already known to have a z value of 1 and stores
 // the result in (x3, y3, z3).  That is to say (x3, y3, z3) = 2*(x1, y1, 1).  It
@@ -597,24 +517,6 @@ func (curve *KoblitzCurve) doubleJacobian(x1, y1, z1, x3, y3, z3 *fieldVal) {
 	// Fall back to generic point doubling which works with arbitrary z
 	// values.
 	curve.doubleGeneric(x1, y1, z1, x3, y3, z3)
-}
-
-// Double returns 2*(x1,y1). Part of the elliptic.Curve interface.
-func (curve *KoblitzCurve) Double(x1, y1 *big.Int) (*big.Int, *big.Int) {
-	if y1.Sign() == 0 {
-		return new(big.Int), new(big.Int)
-	}
-
-	// Convert the affine coordinates from big integers to field values
-	// and do the point doubling in Jacobian projective space.
-	fx1, fy1 := curve.bigAffineToField(x1, y1)
-	fx3, fy3, fz3 := new(fieldVal), new(fieldVal), new(fieldVal)
-	fOne := new(fieldVal).SetInt(1)
-	curve.doubleJacobian(fx1, fy1, fOne, fx3, fy3, fz3)
-
-	// Convert the Jacobian coordinate field values back to affine big
-	// integers.
-	return curve.fieldJacobianToBigAffine(fx3, fy3, fz3)
 }
 
 // splitK returns a balanced length-two representation of k and their signs.
@@ -753,205 +655,8 @@ func naf(k []byte) ([]byte, []byte) {
 	return retPos[1:], retNeg[1:]
 }
 
-// ScalarMult returns k*(Bx, By) where k is a big endian integer.
-// Part of the elliptic.Curve interface.
-func (curve *KoblitzCurve) ScalarMult(Bx, By *big.Int, k []byte) (*big.Int, *big.Int) {
-	// Point Q = ∞ (point at infinity).
-	qx, qy, qz := new(fieldVal), new(fieldVal), new(fieldVal)
-
-	// Decompose K into k1 and k2 in order to halve the number of EC ops.
-	// See Algorithm 3.74 in [GECC].
-	k1, k2, signK1, signK2 := curve.splitK(curve.moduloReduce(k))
-
-	// The main equation here to remember is:
-	//   k * P = k1 * P + k2 * ϕ(P)
-	//
-	// P1 below is P in the equation, P2 below is ϕ(P) in the equation
-	p1x, p1y := curve.bigAffineToField(Bx, By)
-	p1yNeg := new(fieldVal).NegateVal(p1y, 1)
-	p1z := new(fieldVal).SetInt(1)
-
-	// NOTE: ϕ(x,y) = (βx,y).  The Jacobian z coordinate is 1, so this math
-	// goes through.
-	p2x := new(fieldVal).Mul2(p1x, curve.beta)
-	p2y := new(fieldVal).Set(p1y)
-	p2yNeg := new(fieldVal).NegateVal(p2y, 1)
-	p2z := new(fieldVal).SetInt(1)
-
-	// Flip the positive and negative values of the points as needed
-	// depending on the signs of k1 and k2.  As mentioned in the equation
-	// above, each of k1 and k2 are multiplied by the respective point.
-	// Since -k * P is the same thing as k * -P, and the group law for
-	// elliptic curves states that P(x, y) = -P(x, -y), it's faster and
-	// simplifies the code to just make the point negative.
-	if signK1 == -1 {
-		p1y, p1yNeg = p1yNeg, p1y
-	}
-	if signK2 == -1 {
-		p2y, p2yNeg = p2yNeg, p2y
-	}
-
-	// NAF versions of k1 and k2 should have a lot more zeros.
-	//
-	// The Pos version of the bytes contain the +1s and the Neg versions
-	// contain the -1s.
-	k1PosNAF, k1NegNAF := naf(k1)
-	k2PosNAF, k2NegNAF := naf(k2)
-	k1Len := len(k1PosNAF)
-	k2Len := len(k2PosNAF)
-
-	m := k1Len
-	if m < k2Len {
-		m = k2Len
-	}
-
-	// Add left-to-right using the NAF optimization.  See algorithm 3.77
-	// from [GECC].  This should be faster overall since there will be a lot
-	// more instances of 0, hence reducing the number of Jacobian additions
-	// at the cost of 1 possible extra doubling.
-	var k1BytePos, k1ByteNeg, k2BytePos, k2ByteNeg byte
-	for i := 0; i < m; i++ {
-		// Since we're going left-to-right, pad the front with 0s.
-		if i < m-k1Len {
-			k1BytePos = 0
-			k1ByteNeg = 0
-		} else {
-			k1BytePos = k1PosNAF[i-(m-k1Len)]
-			k1ByteNeg = k1NegNAF[i-(m-k1Len)]
-		}
-		if i < m-k2Len {
-			k2BytePos = 0
-			k2ByteNeg = 0
-		} else {
-			k2BytePos = k2PosNAF[i-(m-k2Len)]
-			k2ByteNeg = k2NegNAF[i-(m-k2Len)]
-		}
-
-		for j := 7; j >= 0; j-- {
-			// Q = 2 * Q
-			curve.doubleJacobian(qx, qy, qz, qx, qy, qz)
-
-			if k1BytePos&0x80 == 0x80 {
-				curve.addJacobian(qx, qy, qz, p1x, p1y, p1z,
-					qx, qy, qz)
-			} else if k1ByteNeg&0x80 == 0x80 {
-				curve.addJacobian(qx, qy, qz, p1x, p1yNeg, p1z,
-					qx, qy, qz)
-			}
-
-			if k2BytePos&0x80 == 0x80 {
-				curve.addJacobian(qx, qy, qz, p2x, p2y, p2z,
-					qx, qy, qz)
-			} else if k2ByteNeg&0x80 == 0x80 {
-				curve.addJacobian(qx, qy, qz, p2x, p2yNeg, p2z,
-					qx, qy, qz)
-			}
-			k1BytePos <<= 1
-			k1ByteNeg <<= 1
-			k2BytePos <<= 1
-			k2ByteNeg <<= 1
-		}
-	}
-
-	// Convert the Jacobian coordinate field values back to affine big.Ints.
-	return curve.fieldJacobianToBigAffine(qx, qy, qz)
-}
-
-// ScalarBaseMult returns k*G where G is the base point of the group and k is a
-// big endian integer.
-// Part of the elliptic.Curve interface.
-func (curve *KoblitzCurve) ScalarBaseMult(k []byte) (*big.Int, *big.Int) {
-	newK := curve.moduloReduce(k)
-	diff := len(curve.bytePoints) - len(newK)
-
-	// Point Q = ∞ (point at infinity).
-	qx, qy, qz := new(fieldVal), new(fieldVal), new(fieldVal)
-
-	// curve.bytePoints has all 256 byte points for each 8-bit window. The
-	// strategy is to add up the byte points. This is best understood by
-	// expressing k in base-256 which it already sort of is.
-	// Each "digit" in the 8-bit window can be looked up using bytePoints
-	// and added together.
-	for i, byteVal := range newK {
-		p := curve.bytePoints[diff+i][byteVal]
-		curve.addJacobian(qx, qy, qz, &p[0], &p[1], &p[2], qx, qy, qz)
-	}
-	return curve.fieldJacobianToBigAffine(qx, qy, qz)
-}
-
 // QPlus1Div4 returns the Q+1/4 constant for the curve for use in calculating
 // square roots via exponention.
 func (curve *KoblitzCurve) QPlus1Div4() *big.Int {
 	return curve.q
-}
-
-var initonce sync.Once
-var secp256k1 KoblitzCurve
-
-func initAll() {
-	initS256()
-}
-
-// fromHex converts the passed hex string into a big integer pointer and will
-// panic is there is an error.  This is only provided for the hard-coded
-// constants so errors in the source code can bet detected. It will only (and
-// must only) be called for initialization purposes.
-func fromHex(s string) *big.Int {
-	r, ok := new(big.Int).SetString(s, 16)
-	if !ok {
-		panic("invalid hex in source file: " + s)
-	}
-	return r
-}
-
-func initS256() {
-	// Curve parameters taken from [SECG] section 2.4.1.
-	secp256k1.CurveParams = new(elliptic.CurveParams)
-	secp256k1.P = fromHex("FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFEFFFFFC2F")
-	secp256k1.N = fromHex("FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFEBAAEDCE6AF48A03BBFD25E8CD0364141")
-	secp256k1.B = fromHex("0000000000000000000000000000000000000000000000000000000000000007")
-	secp256k1.Gx = fromHex("79BE667EF9DCBBAC55A06295CE870B07029BFCDB2DCE28D959F2815B16F81798")
-	secp256k1.Gy = fromHex("483ADA7726A3C4655DA4FBFC0E1108A8FD17B448A68554199C47D08FFB10D4B8")
-	secp256k1.BitSize = 256
-	secp256k1.H = 1
-	secp256k1.q = new(big.Int).Div(new(big.Int).Add(secp256k1.P,
-		big.NewInt(1)), big.NewInt(4))
-
-	// Provided for convenience since this gets computed repeatedly.
-	secp256k1.byteSize = secp256k1.BitSize / 8
-
-	// Deserialize and set the pre-computed table used to accelerate scalar
-	// base multiplication.  This is hard-coded data, so any errors are
-	// panics because it means something is wrong in the source code.
-	if err := loadS256BytePoints(); err != nil {
-		panic(err)
-	}
-
-	// Next 6 constants are from Hal Finney's bitcointalk.org post:
-	// https://bitcointalk.org/index.php?topic=3238.msg45565#msg45565
-	// May he rest in peace.
-	//
-	// They have also been independently derived from the code in the
-	// EndomorphismVectors function in gensecp256k1.go.
-	secp256k1.lambda = fromHex("5363AD4CC05C30E0A5261C028812645A122E22EA20816678DF02967C1B23BD72")
-	secp256k1.beta = new(fieldVal).SetHex("7AE96A2B657C07106E64479EAC3434E99CF0497512F58995C1396C28719501EE")
-	secp256k1.a1 = fromHex("3086D221A7D46BCDE86C90E49284EB15")
-	secp256k1.b1 = fromHex("-E4437ED6010E88286F547FA90ABFE4C3")
-	secp256k1.a2 = fromHex("114CA50F7A8E2F3F657C1108D9D44CFD8")
-	secp256k1.b2 = fromHex("3086D221A7D46BCDE86C90E49284EB15")
-
-	// Alternatively, we can use the parameters below, however, they seem
-	//  to be about 8% slower.
-	// secp256k1.lambda = fromHex("AC9C52B33FA3CF1F5AD9E3FD77ED9BA4A880B9FC8EC739C2E0CFC810B51283CE")
-	// secp256k1.beta = new(fieldVal).SetHex("851695D49A83F8EF919BB86153CBCB16630FB68AED0A766A3EC693D68E6AFA40")
-	// secp256k1.a1 = fromHex("E4437ED6010E88286F547FA90ABFE4C3")
-	// secp256k1.b1 = fromHex("-3086D221A7D46BCDE86C90E49284EB15")
-	// secp256k1.a2 = fromHex("3086D221A7D46BCDE86C90E49284EB15")
-	// secp256k1.b2 = fromHex("114CA50F7A8E2F3F657C1108D9D44CFD8")
-}
-
-// S256 returns a Curve which implements secp256k1.
-func S256() *KoblitzCurve {
-	initonce.Do(initAll)
-	return &secp256k1
 }
